+++ dangerfile.ts
@@ -70,6 +70,7 @@ const testFilesThatDontExist = correspondingTestsForAppFiles
   .filter(f => !f.includes("Routes")) // skip routes, kinda untestable
   .filter(f => !f.includes("NativeModules")) // skip modules that are native, they are untestable
   .filter(f => !f.includes("lib/relay/")) // skip modules that are native, they are untestable
+  .filter(f => !f.includes("Storybooks/")) // skip modules that are native, they are untestable
   .filter(f => !fs.existsSync(f))
 
 if (testFilesThatDontExist.length > 0) {